scripts/requirements.txt
psycopg2-binary
numpy
matplotlib
seaborn
python-dotenv
